# packages/proto/gen/openapi/rioku/v1/api.full.json
{
  "components": {
    "parameters": {
      "ID": {
        "description": "Resource id (UUID).",
        "in": "path",
        "name": "id",
        "required": true,
        "schema": {
          "format": "uuid",
          "type": "string"
        }
      },
      "Tenant": {
        "description": "Tenant slug (e.g. `default`, `acme`).",
        "in": "path",
        "name": "tenant",
        "required": true,
        "schema": {
          "pattern": "^[a-z0-9-]+$",
          "type": "string"
        }
      }
    },
    "schemas": {
      "Link": {
        "description": "OpenAPI/HAL hypermedia reference. Every resource emits a\n`_links` map keyed by relation name (`self`, `owner`, `tools`,\n`traces`, etc.) with a Link object.\n",
        "properties": {
          "href": {
            "format": "uri-reference",
            "type": "string"
          },
          "templated": {
            "type": "boolean"
          }
        },
        "required": [
          "href"
        ],
        "type": "object"
      },
      "LinkSet": {
        "additionalProperties": {
          "$ref": "#/components/schemas/Link"
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerToken": {
        "bearerFormat": "JWT",
        "description": "API key or short-lived JWT issued by /api/v1/auth/token.",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "description": "Web-UI session cookie issued by /api/v1/auth/login.",
        "in": "cookie",
        "name": "rioku_sid",
        "type": "apiKey"
      }
    }
  },
  "consumes": [
    "application/json"
  ],
...
  "info": {
    "contact": {
      "name": "Rioku Labs",
      "url": "https://github.com/riokulabs/rioku"
    },
    "description": "Rioku is an open-source, AI-native API gateway platform built on\nCaddy. This document is the canonical OpenAPI 3.1 specification\nfor every REST endpoint exposed by the daemon.\n\nConventions:\n\n- Tenant-scoped paths live under `/api/v1/t/{tenant}/...`.\n- Every resource exposes `OPTIONS` for capability discovery.\n- Both `PUT` (replace, full body) and `PATCH` (partial, RFC 7396\n  merge-patch) are supported on every resource that takes\n  mutations.\n- Relations carry both the foreign-key id AND a `_links.\u003crelation\u003e.href`\n  pointing at the canonical resource path.\n- Pagination is cursor-based (`?page_token=`); page size is\n  bounded at 1000.\n",
    "license": {
      "name": "BUSL-1.1",
      "url": "https://github.com/riokulabs/rioku/blob/main/LICENSE"
    },
    "title": "Rioku API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/t/{tenant}/routes": {
      "get": {
        "operationId": "listRoutes",
        "responses": {
          "200": {
            "description": "ok"
          }
        },
        "summary": "List routes",
        "tags": [
          "routes"
        ]
      },
      "options": {
        "responses": {
          "204": {
            "description": "no content"
          }
        }
      },
      "parameters": [
        {
          "$ref": "#/components/parameters/Tenant"
        }
      ],
      "post": {
        "operationId": "createRoute",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/v1Route"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "created"
          }
        },
        "summary": "Create a route",
        "tags": [
          "routes"
        ]
      }
    },
...
    "/api/v1/t/{tenant}/services": {
      "get": {
        "operationId": "listServices",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "_links": {
                      "$ref": "#/components/schemas/LinkSet"
                    },
                    "items": {
                      "items": {
                        "$ref": "#/components/schemas/v1Service"
                      },
                      "type": "array"
                    },
                    "total": {
                      "type": "integer"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "paginated services collection"
          }
        },
        "summary": "List services",
        "tags": [
          "services"
        ]
      },
      "options": {
        "responses": {
          "204": {
            "description": "no content"
          }
        },
        "summary": "Discover supported methods",
        "tags": [
          "services"
        ]
      },
      "parameters": [
        {
          "$ref": "#/components/parameters/Tenant"
        }
      ],
      "post": {
        "operationId": "createService",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/v1Service"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "service created"
          }
        },
        "summary": "Create a service",
        "tags": [
          "services"
        ]
      }
    },
    "/api/v1/t/{tenant}/services/{id}": {
      "delete": {
        "operationId": "deleteService",
        "responses": {
          "204": {
            "description": "deleted"
          }
        },
        "summary": "Delete service",
        "tags": [
          "services"
        ]
      },
      "get": {
        "operationId": "getService",
        "responses": {
          "200": {
            "description": "ok"
          }
        },
        "summary": "Get service",
        "tags": [
          "services"
        ]
      },
      "options": {
        "responses": {
          "204": {
            "description": "no content"
          }
        },
        "tags": [
          "services"
        ]
      },
      "parameters": [
        {
          "$ref": "#/components/parameters/Tenant"
        },
        {
          "$ref": "#/components/parameters/ID"
        }
      ],
      "patch": {
        "operationId": "patchService",
        "requestBody": {
          "content": {
            "application/merge-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/v1Service"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "ok"
          }
        },
        "summary": "Patch service (RFC 7396 merge-patch)",
        "tags": [
          "services"
        ]
      },
      "put": {
        "operationId": "replaceService",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/v1Service"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "ok"
          }
        },
        "summary": "Replace service",
        "tags": [
          "services"
        ]
      }
    },
    "/api/v1/t/{tenant}/services/{id}/force-reload": {
      "options": {
        "responses": {
          "204": {
            "description": "no content"
          }
        },
        "tags": [
          "services"
        ]
      },
      "parameters": [
        {
          "$ref": "#/components/parameters/Tenant"
        },
        {
          "$ref": "#/components/parameters/ID"
        }
      ],
      "post": {
        "operationId": "forceReloadService",
        "responses": {
          "202": {
            "description": "reload queued"
          }
        },
        "summary": "Force-reload this service's Caddy admin block",
        "tags": [
          "services"
        ]
      }
    },
    "/api/v1/t/{tenant}/services/{id}/routes": {
      "get": {
        "operationId": "listRoutesByService",
        "responses": {
          "200": {
            "description": "paginated route collection"
          }
        },
        "summary": "List routes targeting this service",
        "tags": [
          "services",
          "routes"
        ]
      },
      "options": {
        "responses": {
          "204": {
            "description": "no content"
          }
        }
      },
      "parameters": [
        {
          "$ref": "#/components/parameters/Tenant"
        },
        {
          "$ref": "#/components/parameters/ID"
        }
      ]
    },
...
  },
  "produces": [
    "application/json"
  ],
  "security": [
    {
      "sessionCookie": []
    },
    {
      "bearerToken": []
    }
  ],
  "servers": [
    {
      "description": "Rioku daemon REST endpoint",
      "url": "{scheme}://{host}/api/v1",
      "variables": {
        "host": {
          "default": "localhost:7778"
        },
        "scheme": {
          "default": "https",
          "enum": [
            "https",
            "http"
          ]
        }
      }
    }
  ],
  "swagger": "2.0",
...
}